fix: Filter out input spikes from wrong layers
Prevent that input spikes which are also returned as events from
the step function stay in the input forever but instead are filtered
out before passing them as input into next scan of step function. Had
to fix test where badly constructed Spike with different dimesions
of idx and current was used.

Change-Id: I79eae98c25244ca30b0ccfb4a4c4d3ec98c359f0