feat: Make event serial more general and composable
Serial function now supports nesting. Also it was generalized to work
for both normal LIFs as well as HardwareLIFs by adding an extra
parameter for general external information. Moreover, it supports
updating of weights during the forward pass and passing a state between
the layers (already used for keeping track of layer_index). At this
point when composing different serials weights/recordings are nested.
This complicates a few things and also would cause problems with the
currently implemented training process. In the next commit it is changed
to force the weights to be flat. More comments and adjustments to type
hints will be made in later commits.

Change-Id: Icd00907b964b90596a1ca30493d9ad6a16af9b57