feat: Vmap apply functions instead of the loss
Changing from vmaping loss_wrapper in batch_wrapper to vmapping
the apply function as well as the ones that calculate the loss.
(And including option to not batch as a function parameter)
Also moving the indexing operation for t_first_spike into the
function reduces the time for the first epoch by half (~20s to
~10s for example for me for analytical) and also makes epochs
after that slightly faster. Must be related to gradient calculation
or tracing somehow.

Change-Id: Id17708b602999034cf797c64a1910a6fe0455a04

Change-Id: Ic672df9ab20b1f2e21056a520ad4a9a182a36af9